+++ core/contract/src/test/java/org/eclipse/dataspaceconnector/contract/policy/PolicyArchiveImplTest.java
@@ -16,6 +16,7 @@ package org.eclipse.dataspaceconnector.contract.policy;
 
 import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.spi.contract.negotiation.store.ContractNegotiationStore;
+import org.eclipse.dataspaceconnector.spi.policy.store.PolicyStore;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.agreement.ContractAgreement;
 import org.junit.jupiter.api.Test;
 
@@ -26,13 +27,15 @@ import static org.mockito.Mockito.when;
 class PolicyArchiveImplTest {
 
     private final ContractNegotiationStore contractNegotiationStore = mock(ContractNegotiationStore.class);
-    private final PolicyArchiveImpl policyArchive = new PolicyArchiveImpl(contractNegotiationStore);
+    private final PolicyStore policyStore = mock(PolicyStore.class);
+    private final PolicyArchiveImpl policyArchive = new PolicyArchiveImpl(contractNegotiationStore, policyStore);
 
     @Test
     void shouldGetPolicyFromAgreement() {
         var policy = Policy.Builder.newInstance().id("policyId").build();
-        var contractAgreement = createContractAgreement(policy);
+        var contractAgreement = createContractAgreement(policy.getUid());
         when(contractNegotiationStore.findContractAgreement("contractId")).thenReturn(contractAgreement);
+        when(policyStore.findById("policyId")).thenReturn(policy);
 
         var result = policyArchive.findPolicyForContract("contractId");
 
@@ -48,13 +51,24 @@ class PolicyArchiveImplTest {
         assertThat(result).isNull();
     }
 
-    private ContractAgreement createContractAgreement(Policy policy) {
+    @Test
+    void shouldReturnNullIfPolicyDoesNotExist() {
+        var contractAgreement = createContractAgreement("policyId");
+        when(contractNegotiationStore.findContractAgreement("contractId")).thenReturn(contractAgreement);
+        when(policyStore.findById("policyId")).thenReturn(null);
+
+        var result = policyArchive.findPolicyForContract("contractId");
+
+        assertThat(result).isNull();
+    }
+
+    private ContractAgreement createContractAgreement(String policyId) {
         return ContractAgreement.Builder.newInstance()
                 .id("any")
                 .consumerAgentId("any")
                 .providerAgentId("any")
                 .assetId("any")
-                .policy(policy)
+                .policyId(policyId)
                 .build();
     }
 }